[Improvement] Split configure and move test cases
