+++ tests/SlowTests.hs
@@ -34,14 +34,14 @@ properGenTests =
     "proper generation tests"
     [ localOption (mkTimeout 180_000_000) $
         testCase "Repair ThreeFixes w/ randomness" $ do
-          let toFix = "tests/ThreeFixes.hs"
+          let toFix = "tests/cases/ThreeFixes.hs"
               repair_target = Nothing
               expected =
                 map
                   unlines
-                  [ [ "diff --git a/tests/ThreeFixes.hs b/tests/ThreeFixes.hs",
-                      "--- a/tests/ThreeFixes.hs",
-                      "+++ b/tests/ThreeFixes.hs",
+                  [ [ "diff --git a/tests/cases/ThreeFixes.hs b/tests/cases/ThreeFixes.hs",
+                      "--- a/tests/cases/ThreeFixes.hs",
+                      "+++ b/tests/cases/ThreeFixes.hs",
                       "@@ -20,1 +20,1 @@ brokenPair = (1, 2, 3)",
                       "-brokenPair = (1, 2, 3)",
                       "+brokenPair = (3, 4, 5)"
@@ -61,14 +61,14 @@ genTests =
     "Generation tests"
     [ localOption (mkTimeout 120_000_000) $
         testCase "Repair TwoFixes" $ do
-          let toFix = "tests/TwoFixes.hs"
+          let toFix = "tests/cases/TwoFixes.hs"
               repair_target = Nothing
               expected =
                 map
                   unlines
-                  [ [ "diff --git a/tests/TwoFixes.hs b/tests/TwoFixes.hs",
-                      "--- a/tests/TwoFixes.hs",
-                      "+++ b/tests/TwoFixes.hs",
+                  [ [ "diff --git a/tests/cases/TwoFixes.hs b/tests/cases/TwoFixes.hs",
+                      "--- a/tests/cases/TwoFixes.hs",
+                      "+++ b/tests/cases/TwoFixes.hs",
                       "@@ -12,1 +12,1 @@ brokenPair = (1, 2)",
                       "-brokenPair = (1, 2)",
                       "+brokenPair = (3, 4)"
@@ -82,14 +82,14 @@ genTests =
           fixDiffs @?= expected,
       localOption (mkTimeout 75_000_000) $
         testCase "Repair ThreeFixes" $ do
-          let toFix = "tests/ThreeFixes.hs"
+          let toFix = "tests/cases/ThreeFixes.hs"
               repair_target = Nothing
               expected =
                 map
                   unlines
-                  [ [ "diff --git a/tests/ThreeFixes.hs b/tests/ThreeFixes.hs",
-                      "--- a/tests/ThreeFixes.hs",
-                      "+++ b/tests/ThreeFixes.hs",
+                  [ [ "diff --git a/tests/cases/ThreeFixes.hs b/tests/cases/ThreeFixes.hs",
+                      "--- a/tests/cases/ThreeFixes.hs",
+                      "+++ b/tests/cases/ThreeFixes.hs",
                       "@@ -20,1 +20,1 @@ brokenPair = (1, 2, 3)",
                       "-brokenPair = (1, 2, 3)",
                       "+brokenPair = (3, 4, 5)"
@@ -103,14 +103,14 @@ genTests =
           fixDiffs @?= expected,
       localOption (mkTimeout 90_000_000) $
         testCase "Repair FourFixes" $ do
-          let toFix = "tests/FourFixes.hs"
+          let toFix = "tests/cases/FourFixes.hs"
               repair_target = Nothing
               expected =
                 map
                   unlines
-                  [ [ "diff --git a/tests/FourFixes.hs b/tests/FourFixes.hs",
-                      "--- a/tests/FourFixes.hs",
-                      "+++ b/tests/FourFixes.hs",
+                  [ [ "diff --git a/tests/cases/FourFixes.hs b/tests/cases/FourFixes.hs",
+                      "--- a/tests/cases/FourFixes.hs",
+                      "+++ b/tests/cases/FourFixes.hs",
                       "@@ -24,1 +24,1 @@ brokenPair = (1, 2, 3, 4)",
                       "-brokenPair = (1, 2, 3, 4)",
                       "+brokenPair = (3, 4, 5, 6)"
