+++ tests/cases/BrokenGCD.hs
