+++ tests/cases/BrokenModule.hs
