+++ tests/cases/FourFixes.hs
