+++ tests/cases/MagicConstant.hs
