+++ tests/cases/ThreeFixes.hs
