+++ tests/cases/TwoFixes.hs
